Move deprecated functions into appimage_legacy.h and appimage_legacy.cpp
